ice-restart: fix functionality
by listening to the transport selectedcandidatepairchange event.
This sample has been relying on the somewhat buggy but useful behavior
of the "legacy" iceconnectionstatechange firing a "connected" event even
if the previous state was already connected.

This behavior is no longer supported in the spec iceconnectionstatechange
event. Alternatively listen to the selectedcandidatepairchange event and
determine the availability of new candidates based on that